requirements.txt
crewai
crewai-tools
langchain-openai
streamlit
python-dotenv
